refactor: organize MCP command modules
